+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## [Unreleased]
 
+- Add `Extension.openOptions()` for interacting with the configured options document in a regular
+  tab.
 - Add `Extension.openPopup()` for interacting with the configured popup document in a regular tab.
 - Add `Extension.openDetailsPage()` for enabling and disabling the extension through Chromium's
   management UI.